backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==1.10.12
httpx==0.25.2
pytest==7.4.3
PyJWT==2.8.0
alembic==1.13.0